Making some changes to actions
