+++ client/app/reader/actions/sharedActions.js
@@ -1,0 +1,4 @@
+export const collectAllTags = (documents) => ({
+  type: Constants.COLLECT_ALL_TAGS_FOR_OPTIONS,
+  payload: documents
+});
