+++ client/app/reader/constants.js
@@ -70,7 +70,7 @@ export const SET_PDF_DOCUMENT = 'SET_PDF_DOCUMENT';
 export const CLEAR_PDF_PAGE = 'CLEAR_PDF_PAGE';
 export const ROTATE_PDF_DOCUMENT = 'ROTATE_PDF_DOCUMENT';
 export const CLEAR_PDF_DOCUMENT = 'CLEAR_PDF_DOCUMENT';
-export const GET_DCOUMENT_TEXT = 'GET_DCOUMENT_TEXT';
+export const GET_DOCUMENT_TEXT = 'GET_DOCUMENT_TEXT';
 export const SET_DOCUMENT_SEARCH = 'SET_DOCUMENT_SEARCH';
 
 export const SET_OPENED_ACCORDION_SECTIONS = 'SET_OPENED_ACCORDION_SECTIONS';
